+++ src/spaceship.h
@@ -1,7 +1,7 @@
 extern float spaceship_altitude;
 extern bool spaceship_flying;
 extern double spaceship_height;
-extern double spaceship_radius;
+extern float spaceship_radius;
 
 void spaceship(void);
 void spaceshipLightsTimer(int value);
